da/requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
six
sqlalchemy
pg8000
cloud-sql-python-connector